+++ PC_ROS_SDK/readme.md
@@ -1,0 +1,55 @@
+## 编译方法
+
+使用catkin编译，执行如下操作
+
+```sh
+catkin_make
+
+```
+
+## 运行方法
+
+```sh
+source devel/setup.bash
+roslaunch ldlidar ld19.launch 
+
+or
+
+rosrun ldlidar ldlidar 
+```
+
+rviz的配置在rviz文件夹下面。
+
+
+## 测试
+
+代码在ubuntun16.04 kinetic版本下测试，使用rviz可视化。
+
+---
+
+## Compiling method
+
+Compile with **catkin** and do the following
+
+```sh
+catkin_make
+
+```
+
+## Run method
+
+```sh
+source devel/setup.bash
+roslaunch ldlidar ld19.launch 
+
+or
+
+rosrun ldlidar ldlidar 
+```
+
+The configuration for rviz is under the rviz folder.
+
+
+## Test
+
+The code was tested under Ubuntu 16.04 Kinetic version, visualized using RVIZ.
