+++ PC_ROS_SDK/src/ldlidar/launch/ld19.launch
@@ -1,0 +1,6 @@
+<launch>
+
+ <node name="LD19" pkg="ldlidar" type="ldlidar" args="LD19" output="screen" >
+ </node>
+
+</launch>
