+++ SBC_ROS_SDK/src/ldlidar/launch/ld06.launch
@@ -1,0 +1,5 @@
+<launch>
+ <param name="product" type="string" value="LD06" />
+ <node name="LD06" pkg="ldlidar" type="ldlidar" output="screen">
+ </node>
+</launch>
