--- src/ldlidar/launch/ld19.launch
@@ -1,6 +1,0 @@
-<launch>
-
- <node name="LD19" pkg="ldlidar" type="ldlidar" args="LD19" output="screen" >
- </node>
-
-</launch>
